Add python 3.12 support
