+++ .github/workflows/test.yml
@@ -28,12 +28,14 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        python-version: ["3.8", "3.11"]
+        python-version: ["3.8", "3.12"]
         include:
           - os: windows-latest
             python-version: "3.9"
           - os: ubuntu-latest
-            python-version: "pypy-3.8"
+            python-version: "3.11"
+          - os: ubuntu-latest
+            python-version: "pypy-3.9"
           - os: macos-latest
             python-version: "3.10"
     steps:
